backend/requirements.txt
fastapi==0.115.5
uvicorn==0.34.0
pydantic==2.10.5
pydantic[email]
bcrypt==4.2.1
python-multipart==0.0.20
supabase==2.10.0
python-dotenv==1.0.1
sendgrid